pyproject.toml
[project]
name = "baby-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.2"
dependencies = [
    "pip>=25.3",
    "pyaudio>=0.2.14",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "librosa>=0.10.0",
    "requests>=2.31.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.32.0",
    "pandas>=2.2.0",
    "google-cloud-pubsub>=2.19.0",
]
